requirements.txt
python-dotenv==1.0.1
requests==2.32.3
jinja2==3.1.4
google-generativeai==0.8.0
pandas==2.2.2
openpyxl==3.1.5
